removed old choose_table method and fixed tests to use new method
