+++ app/domain/interface.py
@@ -25,9 +25,6 @@ class Actions(abc.ABC):
     def search_catalogs(self, r: model.SearchCatalogsRequest) -> model.SearchCatalogsResponse:
         raise NotImplementedError("not implemented")
 
-    def choose_table(self, r: model.ChooseTableRequest) -> model.ChooseTableResponse:
-        raise NotImplementedError("not implemented")
-
     def start_task(self, r: model.StartTaskRequest) -> model.StartTaskResponse:
         raise NotImplementedError("not implemented")
