+++ app/domain/model/pipeline.py
@@ -33,14 +33,3 @@ class Catalog:
 @dataclass
 class SearchCatalogsResponse:
     catalogs: list[Catalog]
-
-
-@dataclass
-class ChooseTableRequest:
-    catalog_id: str
-    table_id: str
-
-
-@dataclass
-class ChooseTableResponse:
-    id: int
