+++ app/presentation/server/__init__.py
@@ -45,7 +45,6 @@ routes = [
     (HTTPMETHOD_POST, "/api/v1/admin/object", handlers.create_object),
     (HTTPMETHOD_GET, "/api/v1/object/names", handlers.get_object_names),
     (HTTPMETHOD_GET, "/api/v1/pipeline/catalogs", handlers.search_catalogs),
-    (HTTPMETHOD_POST, "/api/v1/pipeline/table", handlers.choose_table),
     (HTTPMETHOD_POST, "/api/v1/task", handlers.start_task),
     (HTTPMETHOD_GET, "/api/v1/task", handlers.get_task_info),
 ]
