+++ app/presentation/server/handlers/__init__.py
@@ -6,7 +6,6 @@ from app.presentation.server.handlers.get_source import get_source
 from app.presentation.server.handlers.get_source_list import get_source_list
 from app.presentation.server.handlers.get_task_info import get_task_info
 from app.presentation.server.handlers.ping import ping
-from app.presentation.server.handlers.pipeline.choose_table import choose_table
 from app.presentation.server.handlers.pipeline.search_catalogs import search_catalogs
 from app.presentation.server.handlers.start_task import start_task
 
@@ -19,7 +18,6 @@ __all__ = [
     "get_source_list",
     "ping",
     "search_catalogs",
-    "choose_table",
     "start_task",
     "get_task_info",
 ]
